requirements.txt
pandas
sqlalchemy
argparse
pyyaml
pymysql
pyodbc
tabulate
pyarrow
fastparquet
